+++ packages/components/organisms/f-checkout/test/component/f-checkout-delivery.component.desktop.spec.js
@@ -15,14 +15,8 @@ describe('f-checkout "delivery" component tests', () => {
         });
 
         it('should enable a user to submit a postcode with correct characters', () => {
-            // Arrange
-            const addressInfo = {
-                postcode: 'AR51 1AA'
-            };
 
             // Act
-            checkout.clearBlurField('addressPostcode');
-            checkout.populateCheckoutForm(addressInfo);
             checkout.goToPayment();
 
             // Assert
@@ -34,7 +28,7 @@ describe('f-checkout "delivery" component tests', () => {
         beforeEach(() => {
             checkout = new Checkout();
             checkout.withQuery('&knob-Service Type', 'delivery')
-                .withQuery('&knob-Is User Logged In', true)
+                .withQuery('&knob-Is User Logged In', false)
                 .withQuery('&knob-Is ASAP available', true)
                 .withQuery('&knob-Locale', 'en-AU');
 
@@ -44,7 +38,6 @@ describe('f-checkout "delivery" component tests', () => {
         it('should prevent more than 50 characters in state field', () => {
             // Arrange
             const field = 'addressAdministrativeArea';
-            checkout.clearCheckoutForm(field);
             const userEntry = 'A'.repeat(50 + 1); // Enter more than allowed
 
             // Act