+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -12,7 +12,7 @@ v3.0.0
 - New colour scheme and border radius in line with Icing Phase 2.
 
 ### Removed
-- Font-family beclaration from the button styles as we have one font for the whole site and there is no need in declaring it on button level.
+- Font-family declaration from the button styles as we have one font for the whole site and there is no need in declaring it on button level.
 
 
 v2.0.0