+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -4,6 +4,28 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v3.0.1
+------------------------------
+*October 5, 2021*
+
+### Changed
+- Icon button loading state fix.
+
+
+v3.0.0
+------------------------------
+*October 5, 2021*
+
+### Changed
+- New colour scheme and border radius in line with Icing Phase 2.
+
+### Removed
+- Font-family declaration from the button styles as we have one font for the whole site and there is no need in declaring it on button level.
+
+### Added
+- `inverse` and `ghostInverse` props for icon buttons
+
+
 v2.0.0
 ------------------------------
 *September 15, 2021*